Add a stand-alone utility to excise proxy-wasm-cpp-host functionality
This commit adds a utility proxy-wasm-cli that is a stand-along
emulation of a proxy server behavior and provides a way to invoke the
ABI functions.

The proxy-wasm-cli utility starts up a GRPC Server. A python client
(proxy_wasm_cli.py) provides a way to talk to the proxm-wasm-cli GRPC
server and execute some simple functions.

Refer to README.md and src/proxy-wasm-cli/README for more details.